+++ src/internet/model/icmpv6-l4-protocol.h
@@ -127,7 +127,7 @@ public:
    * by setting the node in the ICMPv6 stack and adding ICMPv6 factory to
    * IPv6 stack connected to the node.
    */
-  void NotifyNewAggregate ();
+  virtual void NotifyNewAggregate ();
 
   /**
    * \brief Get the protocol number.
@@ -319,7 +319,7 @@ public:
    * \param interface the interface
    * \param addr the IPv6 address
    */
-  void FunctionDadTimeout (Ipv6Interface* interface, Ipv6Address addr);
+  virtual void FunctionDadTimeout (Ipv6Interface* interface, Ipv6Address addr);
 
   /**
    * \brief Lookup in the ND cache for the IPv6 address
@@ -332,7 +332,7 @@ public:
    * \param hardwareDestination hardware address
    * \return true if the address is in the ND cache, the hardwareDestination is updated.
    */
-  bool Lookup (Ipv6Address dst, Ptr<NetDevice> device, Ptr<NdiscCache> cache, Address* hardwareDestination);
+  virtual bool Lookup (Ipv6Address dst, Ptr<NetDevice> device, Ptr<NdiscCache> cache, Address* hardwareDestination);
 
   /**
    * \brief Lookup in the ND cache for the IPv6 address (similar as ARP protocol).
@@ -346,12 +346,12 @@ public:
    * \param hardwareDestination hardware address
    * \return true if the address is in the ND cache, the hardwareDestination is updated.
    */
-  bool Lookup (Ptr<Packet> p, const Ipv6Header & ipHeader, Ipv6Address dst, Ptr<NetDevice> device, Ptr<NdiscCache> cache, Address* hardwareDestination);
+  virtual bool Lookup (Ptr<Packet> p, const Ipv6Header & ipHeader, Ipv6Address dst, Ptr<NetDevice> device, Ptr<NdiscCache> cache, Address* hardwareDestination);
 
   /**
    * \brief Send a Router Solicitation.
    * \param src link-local source address
-   * \param dst destination address (usually ff02::2 i.e all-routers)
+   * \param dst destination address (usually ff02::2 i.e., all-routers)
    * \param hardwareAddress link-layer address (SHOULD be included if src is not ::)
    */
   void SendRS (Ipv6Address src, Ipv6Address dst,  Address hardwareAddress);
@@ -379,6 +379,12 @@ public:
    * \return the number of stream indices assigned by this model
    */
   int64_t AssignStreams (int64_t stream);
+
+  /**
+   * Get the DAD timeout
+   * \return the DAD timeout
+   */
+  Time GetDadTimeout () const;
 
 protected:
   /**
@@ -417,6 +423,14 @@ protected:
    * \param interface the interface from which the packet is coming
    */
   void HandleRS (Ptr<Packet> p, Ipv6Address const &src, Ipv6Address const &dst, Ptr<Ipv6Interface> interface);
+
+  /**
+   * \brief Router Solicitation Timeout handler.
+   * \param src link-local source address
+   * \param dst destination address (usually ff02::2 i.e all-routers)
+   * \param hardwareAddress link-layer address (SHOULD be included if src is not ::)
+   */
+  virtual void HandleRsTimeout (Ipv6Address src, Ipv6Address dst,  Address hardwareAddress);
 
   /**
    * \brief Receive Router Advertisement method.
@@ -523,7 +537,6 @@ protected:
    */
   CacheList m_cacheList;
 
-private:
   /**
    * \brief Neighbor Discovery node constants: max multicast solicitations.
    */
@@ -533,6 +546,43 @@ private:
    * \brief Neighbor Discovery node constants: max unicast solicitations.
    */
   uint8_t m_maxUnicastSolicit;
+
+  /**
+   * \brief Initial multicast RS retransmission time [\RFC{7559}].
+   */
+  Time m_rsInitialRetransmissionTime;
+
+  /**
+   * \brief Maximum time between multicast RS retransmissions [\RFC{7559}]. Zero means unbound.
+   */
+  Time m_rsMaxRetransmissionTime;
+
+  /**
+   * \brief Maximum number of multicast RS retransmissions [\RFC{7559}]. Zero means unbound.
+   */
+  uint32_t m_rsMaxRetransmissionCount;
+
+  /**
+   * \brief Maximum duration of multicast RS retransmissions [\RFC{7559}]. Zero means unbound.
+   */
+  Time m_rsMaxRetransmissionDuration;
+
+  /**
+   * \brief Multicast RS retransmissions counter [\RFC{7559}].
+   *
+   * Zero indicate a first transmission, greater than zero means retranmsisisons.
+   */
+  uint32_t m_rsRetransmissionCount {0};
+
+  /**
+   * \brief Previous multicast RS retransmissions timeout [\RFC{7559}].
+   */
+  Time m_rsPrevRetransmissionTimeout;
+
+  /**
+   * \brief First multicast RS transmissions [\RFC{7559}].
+   */
+  Time m_rsFirstTransmissionTime;
 
   /**
    * \brief Neighbor Discovery node constants: reachable time.
@@ -558,6 +608,21 @@ private:
    * \brief Random jitter before sending solicitations
    */
   Ptr<RandomVariableStream> m_solicitationJitter;
+
+  /**
+   * \brief Random jitter for RS retransmissions
+   */
+  Ptr<UniformRandomVariable> m_rsRetransmissionJitter;
+
+  /**
+   * \brief DAD timeout
+   */
+  Time m_dadTimeout;
+
+  /**
+   * RS timeout handler event
+   */
+  EventId m_handleRsTimeoutEvent;
 
   IpL4Protocol::DownTargetCallback6 m_downTarget; //!< callback to Ipv6::Send