+++ RELEASE_NOTES.md
@@ -18,6 +18,8 @@ Release 3-dev
 - (lr-wpan) Adds PAN descriptor, CommandPayload Header and Capability Field
 - (wifi) !984 - MultiUserScheduler can request channel access periodically
 - (lr-wpan) !991 - Adds MAC ED scan support
+- (internet) !996 - IPv6 Router Solicitations (RS) are now retrnamitted up to 4 times, following RFC 5779.
+
 
 ### Bugs fixed